# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Usuarios",
        "description": "API RESTful para gestionar usuarios del sistema",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/appsalon/public/api",
            "description": "Servidor Local de Desarrollo"
        }
    ],
    "paths": {
...
        "/citas": {
            "get": {
                "tags": [
                    "Citas"
                ],
                "summary": "Obtener todas las citas",
                "operationId": "a02b676c8f00707c7fa99663a7205759",
                "responses": {
                    "200": {
                        "description": "Listado de citas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Citas"
                ],
                "summary": "Crear una nueva cita",
                "operationId": "e4bb96f25e4201c81d29a1b4216d40be",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fecha",
                                    "hora",
                                    "usuario_id"
                                ],
                                "properties": {
                                    "fecha": {
                                        "type": "string"
                                    },
                                    "hora": {
                                        "type": "string"
                                    },
                                    "usuario_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cita creada"
                    }
                }
            }
        },
        "/citas/{id}": {
            "get": {
                "tags": [
                    "Citas"
                ],
                "summary": "Obtener una cita por ID",
                "operationId": "41cd60146caa86605f2b133ba46f6984",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cita encontrada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Citas"
                ],
                "summary": "Actualizar una cita",
                "operationId": "0dc31503e4eb3c1610480caa522916d5",
                "responses": {
                    "200": {
                        "description": "Actualizada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Citas"
                ],
                "summary": "Eliminar cita",
                "operationId": "5d232a3b6be3cc76d1687b619580290b",
                "responses": {
                    "200": {
                        "description": "Eliminada"
                    }
                }
            }
        },
...
        "/tokens": {
            "get": {
                "tags": [
                    "Gestión de Tokens"
                ],
                "summary": "Listar tokens activos",
                "description": "Retorna todos los tokens activos del usuario autenticado",
                "operationId": "listTokens",
                "responses": {
                    "200": {
                        "description": "Lista de tokens obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tokens obtenidos exitosamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "auth_token"
                                                    },
                                                    "last_used_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tokens/{id}": {
            "delete": {
                "tags": [
                    "Gestión de Tokens"
                ],
                "summary": "Revocar token específico",
                "description": "Revoca un token específico del usuario autenticado",
                "operationId": "revokeToken",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del token a revocar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token revocado exitosamente"
                    },
                    "404": {
                        "description": "Token no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}